feat: Implement photo upload and management with S3 integration and add Swagger API documentation.

+++ api/src/main/java/com/AchadosPerdidos/API/Application/Services/PhotoService.java
@@ -17,7 +17,6 @@ import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
-import java.math.BigInteger;
 import java.util.List;
 
 @Service
@@ -145,7 +144,7 @@ public class PhotoService implements IPhotoService {
         Photo photo = new Photo();
         photo.setUrl(url);
         photo.setFileName(file.getOriginalFilename());
-        photo.setSizeBytes(BigInteger.valueOf(file.getSize()));
+        photo.setSizeBytes(file.getSize());
         photo.setFileType(file.getContentType());
         return photo;
     }
